refactor(e2e): split OpenID tests into openid.spec.ts (ICRC-3) and openid-legacy.spec.ts
